Move constants module one level up
